Separate paraid injection to own pallet
