+++ test/parachain/runtime/src/message_example.rs
@@ -18,7 +18,7 @@
 //! downward messages.
 
 use frame_support::{
-	decl_event, decl_module, decl_storage,
+	decl_event, decl_module,
 	traits::{Currency, ExistenceRequirement, WithdrawReason},
 };
 use frame_system::ensure_signed;
@@ -53,19 +53,6 @@ pub trait Trait: frame_system::Trait {
 
 	/// The sender of XCMP messages.
 	type XCMPMessageSender: XCMPMessageSender<XCMPMessage<Self::AccountId, BalanceOf<Self>>>;
-}
-
-// This pallet's storage items.
-decl_storage! {
-	trait Store for Module<T: Trait> as ParachainUpgrade {}
-	add_extra_genesis {
-		config(parachain_id): ParaId;
-		build(|config: &Self| {
-			// This is basically a hack to make the parachain id easily configurable.
-			// Could also be done differently, but yeah..
-			crate::ParachainId::set(&config.parachain_id);
-		});
-	}
 }
 
 decl_event! {
