+++ test/parachain/src/chain_spec.rs
@@ -17,7 +17,7 @@
 use cumulus_primitives::ParaId;
 use parachain_runtime::{
 	AccountId, BalancesConfig, GenesisConfig, Signature, SudoConfig, SystemConfig,
-	TokenDealerConfig, WASM_BINARY,
+	ParaIdConfig, WASM_BINARY,
 };
 use sc_chain_spec::{ChainSpecExtension, ChainSpecGroup};
 use sc_service::ChainType;
@@ -140,6 +140,6 @@ fn testnet_genesis(
 				.collect(),
 		}),
 		pallet_sudo: Some(SudoConfig { key: root_key }),
-		message_example: Some(TokenDealerConfig { parachain_id: id }),
+		para_id: Some(ParaIdConfig { parachain_id: id }),
 	}
 }
